requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Single cell analysis
anndata>=0.8.0
scanpy>=1.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Optional dependencies (install separately if needed)
# scimilarity  # For foundation model
# scvi-tools>=1.0.0  # For scVI batch correction
